JSON files for Tutors/Splash Page
These JSON files contain information to use in the sandbox. The splashScreen.json is for the splash screen for the learning center. The TutorClassCardSelection is for showing tutors with how they tutor, what they look like, what their times are, and if they are ONLINE/OFFLINE. TutorialLandingScreen is for showing what subjects have tutors for.